Create UpgradeCitrixBackend.ps1
This function will detect which Citrix backend server application is installed and upgrade the components by mounting the appropriate ISO file and perform a silent update. If Provisioning Services is present in the environment the PVS iso must also be present on the ISOPath location. ISO names cannot be altered.